Remove `GetFileListAsync` search pattern parameter
The search pattern parameter to the `GetFileListAsync` method was
modeled after the filter parameter that can be provided to the
`System.IO.Directory.GetFiles()` method. This cannot necessarily be
implemented consistently across all possible implementations of an
`IFileSystem`, and so has been removed until a more suitable option is
found.

This change removes this parameter and makes the `GetFileListAsync`
method consistent with all other methods on the `IFileSystem`
interface, as this now takes a `path` parameter.