Refactor solution structure
The physical file system implementation has be separated out into a new
project. This allows the abstractions and the implementation to be
distributed, and therefore consumed, independently. This is useful if
you never need to use the physical file system implementation, but do
want to use the abstractions with another implementation.