Remove file stream seek from physical file save
Previously the `FileSystem.SaveFileAsync()` method would seek to the
begining of the file stream passed to it before saving the file. This
is likely to be surprising to callers, and so we now assume that any
caller has seeked to the correct position to begin reading the file
stream to save from.