Remove directory parameter from File Storage implementation
Previously, the Azure File Storage implementation took both a share name
and a directory name. Rather than scoping file requests to a
sub-directory of the file share, removing this parameter allows this
implementation to target the root directory of the file share. This
change makes the signature of this implementation more consistent with
that of the Blob Storage implementation, which only takes a container
name.

In the future, we can look to scope these classes to sub-directories, in
the case of File Storage, or path prefixes, in the case of Blob Storage.